requirements.txt
asyncio
aiohttp
customtkinter
opencv-python
numpy
onnxruntime-directml
pyautogui
pygetwindow
requests
toml
torch
pillow
discord.py
easyocr
shapely
ultralytics==8.2.31
bettercam
packaging